requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
certifi==2024.6.2
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
numpy==1.26.4
openai==1.33.0
packaging==24.1
pillow==10.3.0
pyautogen==0.2.28
pydantic==2.7.3
pydantic_core==2.18.4
pyee==11.1.0
pyppeteer==2.0.0
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
setuptools==69.5.1
sniffio==1.3.1
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.18
websockets==10.4
wheel==0.43.0
zipp==3.19.2
